requirements.txt
requests
flask
flask-cors
Pillow
opencv-python
rembg